refacto: moved loadGLTFModel function into a util file
